test: split jsx integration tests
